+++ Code/mdsCollection.cxx
@@ -294,6 +294,16 @@ bool Collection::Delete(bool deleteOnDisk)
     }
 
   query.str(std::string());
+  query << "DELETE FROM community2collection WHERE collection_id='" <<
+    m_Collection->GetId() << "'";
+  if(!m_Database->GetDatabase()->ExecuteQuery(query.str().c_str()))
+    {
+    m_Database->GetDatabase()->ExecuteQuery("ROLLBACK");
+    m_Database->GetDatabase()->Close();
+    return false;
+    }
+
+  query.str(std::string());
   query << "DELETE FROM dirty_resource WHERE uuid='" <<
     m_Collection->GetUuid() << "'";
   if(!m_Database->GetDatabase()->ExecuteQuery(query.str().c_str()))
